Altered replaceCallback($matches) to filter output based on the project name and the wpgh_included_projects option. If the project name in $projects->name matches a project name stored in wpgh_included_projects it will display the information. If wpgh_included_projects is empty all projects will be displayed.
Altered registerAdmin(). It will now set wpgh_included_projects option
on form submit. It will also the allowed projects as an array to the $data
variable with the index wpgh_included_projects.

Added getIncludedProjectsArray(). This method will return an array of
the projects that are allowed to be displayed or an empty string if
there are no project names set.

Added cleanIncludedProjectsString($str). This method will make sure the
allowed projects string is properly formatted. Properly formatted is all lower
case and no spaces before or after a comma.